LayoutTests/accessibility/custom-elements/controls-expected.txt

This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS labelForControl(accessibilityController.accessibleElementById("custom-tab-1").ariaControlsElementAtIndex(0)) is "AXValue: Panel 1"
PASS labelForControl(accessibilityController.accessibleElementById("custom-tab-1").ariaControlsElementAtIndex(1)) is "AXValue: Panel 2"
PASS labelForControl(accessibilityController.accessibleElementById("custom-tab-2").ariaControlsElementAtIndex(0)) is "AXValue: Panel 3"
PASS accessibilityController.accessibleElementById("custom-tab-2").ariaControlsElementAtIndex(1) is null
PASS successfullyParsed is true

TEST COMPLETE
Panel 1
Panel 2
Panel 3
Panel 4
